Retire domain v1 primitives
