+++ crates/pallet-domains/src/block_tree.rs
@@ -8,8 +8,7 @@ use codec::{Decode, Encode};
 use frame_support::{ensure, PalletError};
 use scale_info::TypeInfo;
 use sp_core::Get;
-use sp_domains::v2::ExecutionReceipt;
-use sp_domains::{DomainId, OperatorId};
+use sp_domains::{DomainId, ExecutionReceipt, OperatorId};
 use sp_runtime::traits::{CheckedSub, One, Saturating, Zero};
 use sp_std::cmp::Ordering;
 use sp_std::vec::Vec;
