+++ crates/subspace-transaction-pool/src/bundle_validator.rs
@@ -6,8 +6,7 @@ use sc_consensus_subspace::get_chain_constants;
 use sp_api::{HeaderT, ProvideRuntimeApi};
 use sp_blockchain::HeaderMetadata;
 use sp_consensus_subspace::{FarmerPublicKey, SubspaceApi};
-use sp_domains::v2::OpaqueBundle;
-use sp_domains::DomainsApi;
+use sp_domains::{DomainsApi, OpaqueBundle};
 use sp_runtime::generic::BlockId;
 use sp_runtime::traits::{Block as BlockT, NumberFor, Zero};
 use std::collections::{HashSet, VecDeque};
