+++ domains/client/block-preprocessor/src/lib.rs
@@ -28,8 +28,7 @@ use runtime_api::InherentExtrinsicConstructor;
 use sc_client_api::BlockBackend;
 use sp_api::ProvideRuntimeApi;
 use sp_blockchain::HeaderBackend;
-use sp_domains::v2::OpaqueBundles;
-use sp_domains::{DomainId, DomainsApi, DomainsDigestItem, ExtrinsicsRoot};
+use sp_domains::{DomainId, DomainsApi, DomainsDigestItem, ExtrinsicsRoot, OpaqueBundles};
 use sp_runtime::traits::{Block as BlockT, Header as HeaderT, NumberFor};
 use std::borrow::Cow;
 use std::collections::{BTreeMap, VecDeque};
