+++ domains/client/domain-operator/src/domain_block_processor.rs
@@ -15,8 +15,7 @@ use sp_consensus::{BlockOrigin, SyncOracle};
 use sp_core::traits::CodeExecutor;
 use sp_domains::fraud_proof::FraudProof;
 use sp_domains::merkle_tree::MerkleTree;
-use sp_domains::v2::ExecutionReceipt;
-use sp_domains::{DomainId, DomainsApi, ExtrinsicsRoot};
+use sp_domains::{DomainId, DomainsApi, ExecutionReceipt, ExtrinsicsRoot};
 use sp_runtime::traits::{Block as BlockT, CheckedSub, HashFor, Header as HeaderT, One, Zero};
 use sp_runtime::Digest;
 use std::sync::Arc;
