+++ domains/client/domain-operator/src/domain_bundle_proposer.rs
@@ -9,8 +9,7 @@ use sc_transaction_pool_api::InPoolTransaction;
 use sp_api::{HeaderT, NumberFor, ProvideRuntimeApi};
 use sp_block_builder::BlockBuilder;
 use sp_blockchain::{HashAndNumber, HeaderBackend};
-use sp_domains::v2::{BundleHeader, ExecutionReceipt};
-use sp_domains::ProofOfElection;
+use sp_domains::{BundleHeader, ExecutionReceipt, ProofOfElection};
 use sp_runtime::traits::{BlakeTwo256, Block as BlockT, Hash as HashT, One, Saturating, Zero};
 use sp_weights::Weight;
 use std::marker::PhantomData;
