+++ domains/client/domain-operator/src/domain_worker.rs
@@ -7,8 +7,7 @@ use sc_client_api::{BlockBackend, BlockImportNotification, BlockchainEvents};
 use sp_api::{ApiError, BlockT, ProvideRuntimeApi};
 use sp_blockchain::{HashAndNumber, HeaderBackend};
 use sp_core::traits::SpawnEssentialNamed;
-use sp_domains::v2::OpaqueBundle;
-use sp_domains::DomainsApi;
+use sp_domains::{DomainsApi, OpaqueBundle};
 use sp_runtime::traits::{Header as HeaderT, NumberFor, One, Saturating};
 use std::collections::hash_map::Entry;
 use std::collections::HashMap;
