+++ test/subspace-test-service/src/lib.rs
@@ -56,8 +56,7 @@ use sp_consensus_subspace::digests::{CompatibleDigestItem, PreDigest};
 use sp_consensus_subspace::FarmerPublicKey;
 use sp_core::traits::SpawnEssentialNamed;
 use sp_core::H256;
-use sp_domains::v2::OpaqueBundle;
-use sp_domains::{GenerateGenesisStateRoot, GenesisReceiptExtension};
+use sp_domains::{GenerateGenesisStateRoot, GenesisReceiptExtension, OpaqueBundle};
 use sp_externalities::Extensions;
 use sp_inherents::{InherentData, InherentDataProvider};
 use sp_keyring::Sr25519Keyring;
